==> Report/01-Introduction.ipynb <==
## The brief

The brief was to:

* Choose an question to evaluate
* Learn the parameters of the models
* Learn the performance of the models
* Use the results of the models to evaluate the question

We first chose the data that we would work with, as finding a suitable dataset was going to be the main problem. For this we will use a Common Vulnerabilities and Exploits (CVE) database. One of the fields of CVE databases is a description of each CVE entry. We will look into how the nature of the CVEs have changed over time by using topic modelling on the entry descriptions. The Models we will using an LDA model but under different training conditions. As these are unsupervised, it is hard to analyse their performance, however we can look at how they have grouped the terms and evaluate them. To finish we will directly compare the models to see how their results differ. 

## Library requirements


import sys
import subprocess
reqs = subprocess.check_output([sys.executable,'-m','pip','freeze'])
installed_packages = [r.decode().split('==')[0] for r in reqs.split()]
if not ('pandas' in installed_packages): 
    !{sys.executable} -m pip install pandas
if not ('sklearn' in installed_packages): 
    !{sys.executable} -m pip install sklearn
if not ('numpy' in installed_packages): 
    !{sys.executable} -m pip install numpy
if not ('scipy' in installed_packages): 
    !{sys.executable} -m pip install scipy   
if not ('seaborn' in installed_packages): 
    !{sys.executable} -m pip install seaborn 
if not ('matplotlib' in installed_packages): 
    !{sys.executable} -m pip install matplotlib 

import pandas as pd
import urllib.request
import os
import csv

## The data

To obtain this dataset in a convenient format, we will download it and, if necessary, process it into a standard form.

We place data in the raw or processed folder depending on the stage of processing, both of which are in the `data` folder of our root. So our file system will look like this:

* /data
  * /data/raw
      * allitems.csv
  * /data/processed
      * formatted_df.csv
      * indepLDAmodels.pickle
      * topicMap.pickle

### Get the data

if not(os.path.exists('../data/raw')):
    os.mkdir('../data/raw')
if not(os.path.exists('../data/processed')):
    os.mkdir('../data/processed')
if not(os.path.exists('../data/raw/allitems.csv.gz')):
    url = 'https://cve.mitre.org/data/downloads/allitems.csv.gz'
    urllib.request.urlretrieve(url, '../data/raw/allitems.csv.gz')
if not(os.path.exists('../data/raw/allitems.csv')):
    col = ["Name","Status","Description","References","Phase","Votes","Comments"]
    data = pd.read_csv('../data/raw/allitems.csv.gz', names = col, encoding='iso8859_15')[10:]
    data.to_csv('../data/raw/allitems.csv')

df = pd.read_csv("../data/raw/allitems.csv", encoding='iso8859_15').drop(columns = ['Unnamed: 0'])
df
