==> Report/02-Pre-Processing.ipynb <==
## Pre-Processing


#!pip3 install nltk
#!pip3 install gensim
#!pip3 install pyLDAvis

import pandas as pd
import requests
import gensim
import string
import nltk
from gensim.utils import simple_preprocess
from gensim.parsing.preprocessing import STOPWORDS
from nltk.corpus import wordnet as wn
from nltk import word_tokenize, pos_tag
from collections import defaultdict
from nltk.stem import WordNetLemmatizer, SnowballStemmer
from nltk.stem.porter import *
from nltk.corpus import stopwords
from nltk.tokenize import word_tokenize

#nltk.download('popular')
#nltk.download('stopwords')
#nltk.download('punkt')
#nltk.download('wordnet')
#nltk.download('averaged_perceptron_tagger')
#nltk.download('omw-1.4')

df = pd.read_csv("../data/raw/allitems.csv", encoding='iso8859_15').drop(columns = ['Unnamed: 0'])
df

We remove all intances of "\*\*" as these vulnerabilties are highlighted to inform us that they are not valid, resons include that they are reserved for a later vulnerabitlity or are a duplicate of another CVE. We will also remove any CVEs from 2022 becuase we would be making judgements about the year based on 1 month's data which would 

df = df[~df.Description.str.startswith('**')]
df = df[~df.Name.str.contains('2022')]
desc = df['Description'].str.lower()

desc[0:10]

We begin the preprocessing by writing a couple of functions for removing stopwords, punctuation and numbers.

stop_words = stopwords.words('english')

def remove_stop_words(text):
    words = text.split(' ')
    text1 = " ".join([i for i in words if i not in stop_words])
    return text1

def clean_entry(text): 
    delete_dict = {sp_character: '' for sp_character in string.punctuation + string.digits}
    delete_dict[' '] = ' ' 
    table = str.maketrans(delete_dict)
    text1 = text.translate(table)

    return text1

Let's apply these functions to a test string and observe the effect.

test_string = "hi my name is bill, you'll like to eat 10 beans. it's cold ( 2deg ) outside."
test_tokens = word_tokenize(test_string)
print(test_tokens)

test_string1 = remove_stop_words(clean_entry(test_string))
print(word_tokenize(test_string1))
test_string2 = clean_entry(remove_stop_words(test_string))
print(word_tokenize(test_string2))

We see here the correct order to apply our functions is removal of stopwords and then cleaning. This is because words like `"you'll"` are in our stopword list but `"youll"` is not, and thus removing punctuation in the cleaning step results in erroneous output. Now we apply the functions to `desc` to clean it up.

desc1 = desc.apply(remove_stop_words)
desc1 = desc1.apply(clean_entry)

print(desc[0])
print(desc1[0])

Irritatingly, we found out after doing the above that `gensim` has a list of stopwords which is much more comprehensive than the list in `nltk`. The code below is therefore almost a complete repeat of our `remove_stop_words` function. We will use both functions to ensure that we cover the largest range of stop words.

all_stop_words = gensim.parsing.preprocessing.STOPWORDS

def remove_all_stop_words(text):
    words = text.split(' ')
    text1 = " ".join([i for i in words if i not in all_stop_words])
    return text1

desc2 = desc1.apply(remove_all_stop_words)
desc2[0]

Now we can start the lemmatising.

tag_map = defaultdict(lambda : wn.NOUN)
tag_map['J'] = wn.ADJ
tag_map['V'] = wn.VERB
tag_map['R'] = wn.ADV

desc3 = []
for line in desc2:
    tokens = word_tokenize(line)
    lemma_function = WordNetLemmatizer()
    desc3.append([lemma_function.lemmatize(token, tag_map[tag[0]]) for token, tag in pos_tag(tokens)]) 

To finish off we save our preprocessed data.

df['Description'] = desc3
df.to_csv('../data/processed/formatted_df.csv')
